+++ examples/1_convenient/index.tsx
@@ -3,8 +3,8 @@
 
 import { Context, MemoSignal_Factory, StateSignal_Factory } from "jsr:@oazmi/tsignal"
 import { InlineHyperZone, VanillaComponentRender, VanillaFragmentRender, VanillaSVGElementRender } from "../../src/mod.ts"
-import { TsignalHTMLRender } from "../../src/tsignal_base/mod.ts"
-import { ReactiveComponentProps } from "../../src/tsignal_inline/mod.ts"
+import { TsignalHTMLRender } from "../../src/tsignal/mod.ts"
+import { ReactiveComponentProps } from "../../src/tsignal/mod.ts"
 
 
 const
