+++ src/tsignal/jsx.ts
@@ -1,6 +1,6 @@
 import type { HTMLElementUniqueMemberKeys, HTMLEventNames, HTMLTagNames, InlineAttrName, InlineEventName, InlineExecuteFn, InlineExecuteName, InlineMemberName, IntrinsicHTMLElements_Factory, IntrinsicSVGElements_Factory, SVGTagNames } from "../dom_core/mod.ts"
-import type { Accessor, DynamicStylable, MaybeAccessor, TsignalStyleProps } from "../tsignal_base/deps.ts"
-import type { ReactiveDynamicStylable } from "../tsignal_base/styling.ts"
+import type { Accessor, DynamicStylable, MaybeAccessor, TsignalStyleProps } from "./deps.ts"
+import type { ReactiveDynamicStylable } from "./styling.ts"
 import type { AttrValue, EventFn, Props, Stringifiable, StyleProps } from "../typedefs.ts"
