+++ src/tsignal/mod.ts
@@ -1,4 +1,6 @@
+import { ComponentProps } from "./jsx.ts"
 export type { Accessor, MaybeAccessor, TsignalStyleProps } from "./deps.ts"
+export type { IntrinsicElements } from "./jsx.ts"
 export {
 	TsignalComponentRender,
 	TsignalFragmentRender,
@@ -6,4 +8,5 @@ export {
 	TsignalSVGRender
 } from "./renderers.ts"
 export { ReactiveDynamicStylable } from "./styling.ts"
+export type ReactiveComponentProps<P, E extends HTMLElement = HTMLElement> = ComponentProps<P, E>
