+++ src/tsignal/styling.ts
